Remove temporary file after plotting.
Also add a flush after write, but the file close should have the same
effect.  I'm trying to deal with a (presumed) race condition in which
only the standard deviation plots, thinking maybe it had something to
do with a file not fully written.  That hypothesis is a huge stretch.